# axum/openapi.yaml
openapi: 3.1.0
info:
  title: Ogonek API
  description: ''
  contact:
    name: API Support
    email: danila.volkov@noxlovette.com
  license:
    name: ''
  version: 1.6.0
servers:
- url: https://api.ogonek.app
  description: Production server
- url: http://localhost:3000
  description: Local development server
paths:
...
  /api/v1/lessons:
    get:
      tags:
      - Lesson
      summary: Lessons belonging to a user
      operationId: list_lessons
      parameters:
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: per_page
        in: query
        description: Items per page
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: search
        in: query
        description: Search term
        required: false
        schema:
          type: string
      - name: assignee
        in: query
        description: Filter by assignee
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Lessons retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedLessons'
        '401':
          description: Unauthorized
      security:
      - api_key: []
    post:
      tags:
      - Lesson
      summary: Creates a lesson with user defaults specified elsewhere
      operationId: create_lesson
      responses:
        '201':
          description: Lesson created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreationId'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
      security:
      - api_key: []
  /api/v1/lessons/{id}:
    get:
      tags:
      - Lesson
      summary: Fetches lesson by id
      operationId: fetch_lesson
      parameters:
      - name: id
        in: path
        description: Lesson ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Lesson retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonFull'
        '401':
          description: Unauthorized
        '404':
          description: Lesson not found
      security:
      - api_key: []
    delete:
      tags:
      - Lesson
      summary: Deletes lesson
      operationId: delete_lesson
      parameters:
      - name: id
        in: path
        description: Lesson ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Lesson deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Lesson not found
      security:
      - api_key: []
    patch:
      tags:
      - Lesson
      summary: Updates lesson
      operationId: update_lesson
      parameters:
      - name: id
        in: path
        description: Lesson ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonUpdate'
        required: true
      responses:
        '204':
          description: Lesson updated successfully
        '401':
          description: Unauthorized
        '404':
          description: Lesson not found
      security:
      - api_key: []
...
  /api/v1/tasks:
    get:
      tags:
      - Task
      summary: Tasks belonging to a user (through assignment or direct ownership)
      operationId: list_tasks
      parameters:
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: per_page
        in: query
        description: Items per page
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: search
        in: query
        description: Search term
        required: false
        schema:
          type: string
      - name: assignee
        in: query
        description: Filter by assignee
        required: false
        schema:
          type: string
      - name: completed
        in: query
        description: Filter by completion status
        required: false
        schema:
          type: boolean
      - name: priority
        in: query
        description: Filter by priority
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTasks'
        '401':
          description: Unauthorized
      security:
      - api_key: []
    post:
      tags:
      - Task
      summary: Creates a new task
      operationId: create_task
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreationId'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
      security:
      - api_key: []
  /api/v1/tasks/{id}:
    get:
      tags:
      - Task
      summary: One Task
      operationId: fetch_task
      parameters:
      - name: id
        in: path
        description: Task ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Task with files retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskWithFilesResponse'
        '401':
          description: Unauthorized
        '404':
          description: Task not found
      security:
      - api_key: []
    put:
      tags:
      - Task
      summary: Toggles completed/not completed on a task
      operationId: toggle_task
      parameters:
      - name: id
        in: path
        description: Task ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Task with files retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskWithFilesResponse'
        '401':
          description: Unauthorized
        '404':
          description: Task not found
      security:
      - api_key: []
    delete:
      tags:
      - Task
      summary: Deletes a task
      operationId: delete_task
      parameters:
      - name: id
        in: path
        description: Task ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Task not found
      security:
      - api_key: []
    patch:
      tags:
      - Task
      summary: Updates the task
      operationId: update_task
      parameters:
      - name: id
        in: path
        description: Task ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
        required: true
      responses:
        '204':
          description: Task updated successfully
        '401':
          description: Unauthorized
        '404':
          description: Task not found
      security:
      - api_key: []
...
components:
  schemas:
...
    CreationId:
      type: object
      description: Simply contains one string, the created ID
      required:
      - id
      properties:
        id:
          type: string
...
    FileSmall:
      type: object
      required:
      - id
      - name
      - size
      - ownerId
      properties:
        id:
          type: string
        mimeType:
          type:
          - string
          - 'null'
        name:
          type: string
        ownerId:
          type: string
        s3Key:
          type:
          - string
          - 'null'
        size:
          type: integer
          format: int64
...
    LessonFull:
      type: object
      description: Grown-up lesson
      required:
      - id
      - title
      - topic
      - markdown
      - assignee
      - createdBy
      - createdAt
      - updatedAt
      - assigneeName
      properties:
        assignee:
          type: string
        assigneeName:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        id:
          type: string
        markdown:
          type: string
        title:
          type: string
        topic:
          type: string
        updatedAt:
          type: string
          format: date-time
    LessonSmall:
      type: object
      description: Mini-lesson
      required:
      - id
      - title
      - topic
      - assigneeName
      - createdAt
      properties:
        assigneeName:
          type: string
        createdAt:
          type: string
          format: date-time
        id:
          type: string
        seen:
          type:
          - boolean
          - 'null'
        title:
          type: string
        topic:
          type: string
    LessonUpdate:
      type: object
      properties:
        assignee:
          type:
          - string
          - 'null'
        createdBy:
          type:
          - string
          - 'null'
        id:
          type:
          - string
          - 'null'
        markdown:
          type:
          - string
          - 'null'
        title:
          type:
          - string
          - 'null'
        topic:
          type:
          - string
          - 'null'
...
    PaginatedLessons:
      type: object
      required:
      - data
      - total
      - page
      - per_page
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LessonSmall'
        page:
          type: integer
          format: int64
        per_page:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
    PaginatedTasks:
      type: object
      required:
      - data
      - total
      - page
      - per_page
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskSmall'
        page:
          type: integer
          format: int64
        per_page:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
...
    TaskFull:
      type: object
      required:
      - id
      - title
      - markdown
      - priority
      - completed
      - createdAt
      - updatedAt
      - createdBy
      - assignee
      - assigneeName
      properties:
        assignee:
          type: string
        assigneeName:
          type: string
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        dueDate:
          type:
          - string
          - 'null'
          format: date-time
        id:
          type: string
        markdown:
          type: string
        priority:
          type: integer
          format: int32
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
...
    TaskSmall:
      type: object
      required:
      - id
      - title
      - priority
      - completed
      - assigneeName
      properties:
        assigneeName:
          type: string
        completed:
          type: boolean
        dueDate:
          type:
          - string
          - 'null'
          format: date-time
        id:
          type: string
        priority:
          type: integer
          format: int32
        seen:
          type:
          - boolean
          - 'null'
        title:
          type: string
    TaskUpdate:
      type: object
      properties:
        assignee:
          type:
          - string
          - 'null'
        completed:
          type:
          - boolean
          - 'null'
        dueDate:
          type:
          - string
          - 'null'
          format: date-time
        markdown:
          type:
          - string
          - 'null'
        priority:
          type:
          - integer
          - 'null'
          format: int32
        title:
          type:
          - string
          - 'null'
    TaskWithFilesResponse:
      type: object
      required:
      - task
      - files
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileSmall'
        task:
          $ref: '#/components/schemas/TaskFull'
...
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
...